更新 iOS ChangeLog

+++ iOS/samples/AVOSDemo/AVOSDemo/AVFileBasic.m
@@ -76,7 +76,7 @@
                     
                     [self log:[NSString stringWithFormat:@"成功得到图片: %@",[img description]]];
                 }
-            } progressBlock:^(int percentDone) {
+            } progressBlock:^(NSInteger percentDone) {
                 [self log:[NSString stringWithFormat:@"加载进度: %d%%",percentDone]];
             }];
